July 13, 2016 3:10 AM (i'm in taiwan)
Missed the time for the last commit. It was somewhere around half an hour ago. Polished up the CSS and added a div to the HTML. Finished the JS for it and is running find.

FUTURE GOALS:
This is a big one, rewrite the site without angular. Should help with speed.