July 13, 2016 3:14 AM (i'm in taiwan)
Got rid of the old snackbar that was annoying... Note. The controls are also now fixed onto the top of the screen. :D
I'm going to take a break and eat dinner now. bye.

FUTURE GOALS:
This is a big one, rewrite the site without angular. Should help with speed.
Add video support to play.js. Will be part of the pure JS project.