July 18, 2016 7:53 AM (i'm in taiwan)
Did Too Much in One Commit. That should be an acronym. DTMOC. Yay. DT-MOC. Created the entire book section. Waiting on Jamie for the Chinese.

FUTURE GOALS:
This is a big one, rewrite the site without angular. Should help with speed.
Add video support to play.js. Will be part of the pure JS project